Revamped the angle comparing version.
Bug:
The old system was flawed because, for instance, if I want an event between of all swipes to the right, it was impossible to set up because the bounding angles are 270 to 90.

Fix:
The newest version instead takes in direction and allowed angle variance, so an event of trigger of all swipes to the right would be Direction {1,0}, and angle variance 90 degrees.